+++ natsserver/doc.go
@@ -1,2 +1,0 @@
-// Package natsserver is used to embed the NATS server in an application.
-package natsserver