+++ server/nats-server.go
@@ -1,15 +1,14 @@
-package natsserver
+package server
 
 import (
+	"fmt"
 	"log"
-	"os"
 	"time"
 
 	"github.com/nats-io/nats-server/v2/server"
 )
 
-// Options for starting the nat server
-type Options struct {
+type natsServerOptions struct {
 	Port       int
 	HTTPPort   int
 	WSPort     int
@@ -19,13 +18,13 @@ type Options struct {
 	TLSTimeout float64
 }
 
-// StartNatsServer starts a nats server instance. This function will block
-// so should be started with a go routine
-func StartNatsServer(o Options) {
+// newNatsServer creates a new nats server instance
+func newNatsServer(o natsServerOptions) (*server.Server, error) {
 	opts := server.Options{
 		Port:          o.Port,
 		HTTPPort:      o.HTTPPort,
 		Authorization: o.Auth,
+		NoSigs:        true,
 	}
 
 	if o.TLSCert != "" && o.TLSKey != "" {
@@ -44,7 +43,7 @@ func StartNatsServer(o Options) {
 		opts.TLSConfig, err = server.GenTLSConfig(&tc)
 
 		if err != nil {
-			log.Fatal("Error setting up TLS: ", err)
+			return nil, fmt.Errorf("Error setting up TLS: %v", err)
 		}
 	}
 
@@ -59,7 +58,7 @@ func StartNatsServer(o Options) {
 	natsServer, err := server.NewServer(&opts)
 
 	if err != nil {
-		log.Fatal("Error create new Nats server: ", err)
+		return nil, fmt.Errorf("Error create new Nats server: %v", err)
 	}
 
 	authEnabled := "no"
@@ -68,18 +67,12 @@ func StartNatsServer(o Options) {
 		authEnabled = "yes"
 	}
 
-	log.Printf("Starting NATS server, port: %v, http port: %v, auth enabled: %v\n",
+	log.Printf("NATS server, port: %v, http port: %v, auth enabled: %v\n",
 		o.Port, o.HTTPPort, authEnabled)
 
 	if o.WSPort != 0 {
 		log.Printf("NATS server WS enabled on port: %v\n", o.WSPort)
 	}
 
-	natsServer.Start()
-
-	natsServer.WaitForShutdown()
-
-	// should never get here, so exit app
-	log.Fatal("Nats server start returned, this should not happen, exitting app!")
-	os.Exit(-1)
+	return natsServer, nil
 }