+++ server/server.go
@@ -1,24 +1,27 @@
-package simpleiot
+package server
 
 import (
+	"context"
 	"errors"
 	"flag"
 	"fmt"
 	"log"
+	"net"
 	"os"
-	"os/signal"
 	"strconv"
 	"strings"
+	"sync"
 	"syscall"
 	"time"
 
+	"github.com/nats-io/nats-server/v2/server"
 	"github.com/nats-io/nats.go"
+	"github.com/oklog/run"
 	"github.com/simpleiot/simpleiot/api"
 	"github.com/simpleiot/simpleiot/assets/files"
 	"github.com/simpleiot/simpleiot/assets/frontend"
 	"github.com/simpleiot/simpleiot/client"
 	"github.com/simpleiot/simpleiot/data"
-	"github.com/simpleiot/simpleiot/natsserver"
 	"github.com/simpleiot/simpleiot/node"
 	"github.com/simpleiot/simpleiot/particle"
 	"github.com/simpleiot/simpleiot/sim"
@@ -32,6 +35,7 @@ type Options struct {
 	DataDir           string
 	HTTPPort          string
 	DebugHTTP         bool
+	DebugLifecycle    bool
 	DisableAuth       bool
 	NatsServer        string
 	NatsDisableServer bool
@@ -47,37 +51,69 @@ type Options struct {
 	OSVersionField    string
 }
 
-// Siot is used to manage the Siot server
-type Siot struct {
-	dbInst  *store.Db
-	options Options
+// Server represents a SIOT server process
+type Server struct {
+	nc                 *nats.Conn
+	options            Options
+	natsServer         *server.Server
+	chNatsClientClosed chan struct{}
+	chStop             chan struct{}
+
+	// sync stuff
+	startedLock sync.Mutex
+	started     bool
+	wait        []chan struct{}
 }
 
-// NewSiot create new siot instance
-func NewSiot(o Options) *Siot {
-	return &Siot{options: o}
+// NewServer creates a new server
+func NewServer(o Options) (*Server, *nats.Conn, error) {
+	chNatsClientClosed := make(chan struct{})
+
+	// start the server side nats client
+	nc, err := nats.Connect(o.NatsServer,
+		nats.Timeout(10*time.Second),
+		nats.PingInterval(60*5*time.Second),
+		nats.MaxPingsOutstanding(5),
+		nats.ReconnectBufSize(5*1024*1024),
+		nats.SetCustomDialer(&net.Dialer{
+			KeepAlive: -1,
+		}),
+		nats.Token(o.AuthToken),
+		nats.RetryOnFailedConnect(true),
+		nats.MaxReconnects(5),
+		nats.ReconnectWait(time.Second*3),
+		nats.ErrorHandler(func(_ *nats.Conn,
+			sub *nats.Subscription, err error) {
+			log.Printf("NATS server client error, sub: %v, err: %s\n", sub.Subject, err)
+		}),
+		nats.ReconnectHandler(func(_ *nats.Conn) {
+			log.Println("Nats server client reconnect")
+		}),
+		nats.ClosedHandler(func(_ *nats.Conn) {
+			log.Println("Nats server client closed")
+			close(chNatsClientClosed)
+		}),
+	)
+
+	return &Server{
+		nc:                 nc,
+		options:            o,
+		chNatsClientClosed: chNatsClientClosed,
+		chStop:             make(chan struct{}),
+	}, nc, err
 }
 
-// Close the siot server
-func (s *Siot) Close() error {
-	if s.dbInst != nil {
-		s.dbInst.Close()
+// Start the server -- only returns if there is an error
+func (s *Server) Start() error {
+	var g run.Group
+
+	logLS := func(m ...any) {}
+
+	if s.options.DebugLifecycle {
+		logLS = func(m ...any) {
+			log.Println(m...)
+		}
 	}
-	// TODO can add a lot more stuff in here for clean shutdown
-	return nil
-}
-
-// Start Simple IoT data store. The nats connection returned
-// can be used with helper functions in the simpleiot nats package.
-// Note, this function cannot be used directly because we don't
-// checkin the frontend assets for the SIOT web ui. See this
-// example for how you can embed SIOT in your project by adding
-// it as a submodule:
-// https://github.com/simpleiot/custom-application-examples/tree/main/example-1
-func (s *Siot) Start() (*nats.Conn, error) {
-	// =============================================
-	// Start server, default action
-	// =============================================
 
 	o := s.options
 
@@ -93,7 +129,10 @@ func (s *Siot) Start() (*nats.Conn, error) {
 		}
 	}
 
-	natsOptions := natsserver.Options{
+	// ====================================
+	// Nats server
+	// ====================================
+	natsOptions := natsServerOptions{
 		Port:       o.NatsPort,
 		HTTPPort:   o.NatsHTTPPort,
 		WSPort:     o.NatsWSPort,
@@ -104,103 +143,207 @@ func (s *Siot) Start() (*nats.Conn, error) {
 	}
 
 	if !o.NatsDisableServer {
-		go natsserver.StartNatsServer(natsOptions)
+		s.natsServer, err = newNatsServer(natsOptions)
+		if err != nil {
+			return fmt.Errorf("Error setting up nats server: %v", err)
+		}
+
+		g.Add(func() error {
+			s.natsServer.Start()
+			s.natsServer.WaitForShutdown()
+			logLS("LS: Exited: nats server")
+			return fmt.Errorf("NATS server stopped")
+		}, func(err error) {
+			s.natsServer.Shutdown()
+			logLS("LS: Shutdown: nats server")
+		})
 	}
 
+	// ====================================
+	// Monitor Nats server client
+	// ====================================
+	g.Add(func() error {
+		// block until client exits
+		<-s.chNatsClientClosed
+		logLS("LS: Exited: nats client")
+		return errors.New("Nats server client closed")
+	}, func(error) {
+		s.nc.Close()
+		logLS("LS: Shutdown: nats client")
+	})
+
+	// ====================================
+	// SIOT Store
+	// ====================================
 	storeParams := store.Params{
 		Type:      o.StoreType,
 		DataDir:   o.DataDir,
 		AuthToken: o.AuthToken,
 		Server:    o.NatsServer,
 		Key:       auth,
+		Nc:        s.nc,
 	}
 
 	siotStore, err := store.NewStore(storeParams)
 
-	if err != nil {
-		return nil, fmt.Errorf("Error starting store: %v", err)
-	}
+	g.Add(func() error {
+		err := siotStore.Start()
+		logLS("LS: Exited: store")
+		return err
+	}, func(err error) {
+		siotStore.Stop(err)
+		logLS("LS: Shutdown: store")
+	})
 
-	var nc *nats.Conn
+	metricsCtx, metricsCancel := context.WithTimeout(context.Background(),
+		time.Second*10)
 
-	// this is a bit of a hack, but we're not sure when the NATS
-	// server will be started, so try several times
-	for i := 0; i < 10; i++ {
-		// FIXME should we get nc with edgeConnect here?
-		nc, err = siotStore.Connect()
+	g.Add(func() error {
+		err := siotStore.WaitStart(metricsCtx)
 		if err != nil {
-			log.Println("NATS local connect retry: ", i)
-			time.Sleep(500 * time.Millisecond)
-			continue
+			logLS("LS: Exited: node manager")
+			return err
 		}
 
-		break
-	}
+		rootNode, err := client.GetNode(s.nc, "root", "")
 
-	if err != nil {
-		return nil, fmt.Errorf("Error connecting to NATs server: %v", err)
-	}
-
-	if nc == nil {
-		return nil, fmt.Errorf("Timeout connecting to NATs server")
-	}
-
-	nodeManager := node.NewManger(nc, o.AppVersion, o.OSVersionField)
-	err = nodeManager.Init()
-	if err != nil {
-		return nil, fmt.Errorf("Error initializing node manager: %v", err)
-	}
-	go nodeManager.Run()
-
-	rootNode, err := client.GetNode(nc, "root", "")
-
-	if err != nil {
-		log.Println("Error getting root id for metrics: ", err)
-	} else if len(rootNode) == 0 {
-		log.Println("Error getting root node, no data")
-	} else {
+		if err != nil {
+			logLS("LS: Exited: store metrics")
+			return fmt.Errorf("Error getting root id for metrics: %v", err)
+		} else if len(rootNode) == 0 {
+			logLS("LS: Exited: store metrics")
+			return fmt.Errorf("Error getting root node, no data")
+		}
 
 		err = siotStore.StartMetrics(rootNode[0].ID)
-		if err != nil {
-			log.Println("Error starting nats metrics: ", err)
-		}
-	}
+		logLS("LS: Exited: store metrics")
+		return err
+	}, func(err error) {
+		metricsCancel()
+		siotStore.StopMetrics(err)
+		logLS("LS: Shutdown: store metrics")
+	})
 
+	// ====================================
+	// Node client manager
+	// ====================================
+	nodeManager := node.NewManger(s.nc, o.AppVersion, o.OSVersionField)
+	nodeManagerCtx, nodeManagerCancel := context.WithTimeout(context.Background(),
+		time.Second*10)
+
+	g.Add(func() error {
+		err := siotStore.WaitStart(nodeManagerCtx)
+		if err != nil {
+			logLS("LS: Exited: node manager")
+			return err
+		}
+
+		// signal that the server is started
+		s.startedLock.Lock()
+		s.started = true
+		for _, c := range s.wait {
+			close(c)
+		}
+		s.startedLock.Unlock()
+
+		err = nodeManager.Start()
+		logLS("LS: Exited: node manager")
+		return err
+	}, func(err error) {
+		nodeManagerCancel()
+		nodeManager.Stop(err)
+		logLS("LS: Shutdown: node manager")
+	})
+
+	// ====================================
+	// Particle client
 	// FIXME move this to a node, or get rid of it
+	// ====================================
+
 	if o.ParticleAPIKey != "" {
 		go func() {
 			err := particle.PointReader("sample", o.ParticleAPIKey,
 				func(id string, points data.Points) {
-					err := client.SendNodePoints(nc, id, points, false)
+					err := client.SendNodePoints(s.nc, id, points, false)
 					if err != nil {
 						log.Println("Error getting particle sample: ", err)
 					}
 				})
 
 			if err != nil {
-				fmt.Println("Get returned error: ", err)
+				log.Println("Get returned error: ", err)
 			}
 		}()
 	}
 
-	go func() {
-		err = api.Server(api.ServerArgs{
-			Port:       o.HTTPPort,
-			NatsWSPort: o.NatsWSPort,
-			GetAsset:   frontend.Asset,
-			Filesystem: frontend.FileSystem(),
-			Debug:      o.DebugHTTP,
-			JwtAuth:    auth,
-			AuthToken:  o.AuthToken,
-			Nc:         nc,
-		})
+	// ====================================
+	// HTTP API
+	// ====================================
+	httpAPI := api.NewServer(api.ServerArgs{
+		Port:       o.HTTPPort,
+		NatsWSPort: o.NatsWSPort,
+		GetAsset:   frontend.Asset,
+		Filesystem: frontend.FileSystem(),
+		Debug:      o.DebugHTTP,
+		JwtAuth:    auth,
+		AuthToken:  o.AuthToken,
+		Nc:         s.nc,
+	})
 
-		if err != nil {
-			log.Fatal("Error starting SIOT HTTP interface: ", err)
+	g.Add(func() error {
+		err := httpAPI.Start()
+		logLS("LS: Exited: http api")
+		return err
+	}, func(err error) {
+		httpAPI.Stop(err)
+		logLS("LS: Shutdown: http api")
+	})
+
+	// Give us a way to stop the server
+	chShutdown := make(chan struct{})
+	g.Add(func() error {
+		select {
+		case <-s.chStop:
+			logLS("LS: Exited: stop handler")
+			return errors.New("Server stopped")
+		case <-chShutdown:
+			logLS("LS: Exited: stop handler")
+			return nil
 		}
-	}()
+	}, func(_ error) {
+		close(chShutdown)
+		logLS("LS: Shutdown: stop handler")
+	})
 
-	return nc, err
+	// now, run all this stuff
+	return g.Run()
+}
+
+// Stop server
+func (s *Server) Stop(err error) {
+	s.nc.Close()
+	close(s.chStop)
+}
+
+// WaitStart waits for server to start. Clients should wait for this
+// to complete before trying to fetch nodes, etc.
+func (s *Server) WaitStart(ctx context.Context) error {
+	s.startedLock.Lock()
+	if s.started {
+		s.startedLock.Unlock()
+		return nil
+	}
+
+	wait := make(chan struct{})
+	s.wait = append(s.wait, wait)
+	s.startedLock.Unlock()
+
+	select {
+	case <-ctx.Done():
+		return errors.New("server wait timeout or canceled")
+	case <-wait:
+		return nil
+	}
 }
 
 var version = "Development"
@@ -216,6 +359,7 @@ func StartArgs(args []string) error {
 
 	// configuration options
 	flagDebugHTTP := flags.Bool("debugHttp", false, "Dump http requests")
+	flagDebugLifecycle := flags.Bool("debugLifecycle", false, "Debug program lifecycle")
 	flagSim := flags.Bool("sim", false, "Start node simulator")
 	flagDisableAuth := flags.Bool("disableAuth", false, "Disable user auth (used for development)")
 	flagPortal := flags.String("portal", "http://localhost:8080", "Portal URL")
@@ -250,7 +394,8 @@ func StartArgs(args []string) error {
 		fmt.Printf("SimpleIOT %v\n", version)
 		os.Exit(0)
 	}
-	fmt.Printf("SimpleIOT %v\n", version)
+
+	log.Printf("SimpleIOT %v\n", version)
 
 	// set up local database
 	dataDir := os.Getenv("SIOT_DATA")
@@ -577,6 +722,7 @@ func StartArgs(args []string) error {
 		DataDir:           dataDir,
 		HTTPPort:          port,
 		DebugHTTP:         *flagDebugHTTP,
+		DebugLifecycle:    *flagDebugLifecycle,
 		DisableAuth:       *flagDisableAuth,
 		NatsServer:        natsServer,
 		NatsDisableServer: *flagNatsDisableServer,
@@ -592,37 +738,36 @@ func StartArgs(args []string) error {
 		OSVersionField:    osVersionField,
 	}
 
-	siot := NewSiot(o)
+	var g run.Group
 
-	nc, err = siot.Start()
-
-	// this is not used yet
-	_ = nc
+	siot, _, err := NewServer(o)
 
 	if err != nil {
-		log.Fatal("Error starting SIOT store: ", err)
+		siot.Stop(nil)
+		return fmt.Errorf("Error starting server: %v", err)
 	}
 
-	sigs := make(chan os.Signal, 1)
-	done := make(chan bool, 1)
+	g.Add(siot.Start, siot.Stop)
 
-	signal.Notify(sigs, syscall.SIGINT, syscall.SIGTERM)
+	g.Add(run.SignalHandler(context.Background(),
+		syscall.SIGINT, syscall.SIGTERM))
 
-	go func() {
-		sig := <-sigs
-		log.Println()
-		log.Println(sig)
-		done <- true
-	}()
+	// add check to make sure server started
+	chStartCheck := make(chan struct{})
+	ctx, cancel := context.WithTimeout(context.Background(), time.Second*9)
+	g.Add(func() error {
+		err := siot.WaitStart(ctx)
+		if err != nil {
+			return errors.New("Timeout waiting for SIOT to start")
+		}
+		<-chStartCheck
+		return nil
+	}, func(err error) {
+		cancel()
+		close(chStartCheck)
+	})
 
-	log.Println("running ...")
-	<-done
-	// cleanup
-	log.Println("cleaning up ...")
-	siot.Close()
-	log.Println("exiting")
-
-	return nil
+	return g.Run()
 }
 
 func parsePointText(s string) (string, data.Point, error) {