+++ test/doc.go
@@ -1,0 +1,2 @@
+// Package test contains utilities and end-to-end tests that don't fit elsewhere
+package test