+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Bug fix release (PR #369)
 
 - Fix invalid users causes panic in Go code #365
 - implement data.Decode/Encode for converting nodes to user structs #384
+- improve startup/shutdown lifecycle #389
 
 ## [[0.2.0] - 2022-05-31](https://github.com/simpleiot/simpleiot/releases/tag/v0.2.0)